select single query implementation
